consolidate import logic
